Move signal handling
